% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_drift}
\alias{adjust_drift}
\title{Adjust a single track with a mean speed vector.}
\usage{
adjust_drift(track, drift_vector)
}
\arguments{
\item{track}{a single track, i.e. element of tracks or trax object.}

\item{drift_vector}{a drift vector. Its dimensions need to correspond to
the track dimension.}
}
\value{
a matrix.
}
\description{
Adjusts a single track for non-specific drift by subtracting a
user-provided mean speed vector.
}
