% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.tracks.trax}
\alias{as.tracks.trax}
\title{Tracks Objects}
\usage{
\method{as.tracks}{trax}(x, tracks_only = FALSE, ...)
}
\arguments{
\item{x}{an object to be coerced or tested.}

\item{tracks_only}{logical, should the output be a tracks class instance?}

\item{...}{extra arguments, passed to \code{\link[celltrackR]{tracks}}.}
}
\value{
an object of 'trax' or 'tracks' class.
}
\description{
The default method for the 'as.tracks' generic.
}
