% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as_trax.list}
\alias{as_trax.list}
\title{Create a trax object from a list.}
\usage{
\method{as_trax}{list}(x, ...)
}
\arguments{
\item{x}{an object to be coerced.}

\item{...}{extra arguments, passed to \code{\link[celltrackR]{tracks}}.}
}
\value{
an object of 'trax' class.
}
\description{
Converts a list to a trax object. See:
\code{\link[celltrackR]{tracks}} for details.
}
