% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as_trax.tracks}
\alias{as_trax.tracks}
\title{Create a trax object.}
\usage{
\method{as_trax}{tracks}(x, ...)
}
\arguments{
\item{x}{an object to be coerced.}

\item{...}{extra arguments, passed to \code{\link[celltrackR]{tracks}}.}
}
\value{
an object of 'trax' class.
}
\description{
Generates a trax object form a 'tracks' class object.
}
\details{
The 'trax' class inherits most of its methods from the 'tracks'
class but provides enhanced plotting and filtering features.
}
