% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{autocov}
\alias{autocov}
\title{Autocovariance of displacement vectors.}
\usage{
autocov(x, method = c("dot_product", "angle"), aggregate = TRUE, ...)
}
\arguments{
\item{x}{a trax object.}

\item{method}{the method of autocovariance calculation: 'dot_product'
(default) or 'angle', as described in details.}

\item{aggregate}{logical: should the results be aggregated in a step-wise
manner? Done with \code{\link[celltrackR]{aggregate.tracks}}.}

\item{...}{extra arguments passed to
\code{\link[celltrackR]{aggregate.tracks}}.}
}
\value{
a tibble with the results. If aggregate is FALSE, contains sets of
autocovariance stats for each track, otherwise the step-wise
statistic values.
}
\description{
Calculates autocovariance of between the displacement vectors
of for subsequent steps. The results may be aggregated with mean (default),
median or other user-supplied function to
\code{\link[celltrackR]{aggregate.tracks}}.
}
\details{
The autocovariance is calculated as a dot product of the vectors
returned by \code{\link[celltrackR]{overallDot}} or overall angle computed
by \code{\link[celltrackR]{overallAngle}}.
}
