% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bic_position}
\alias{bic_position}
\title{Check if cell motion can be described by multiple Gaussian distributions.}
\usage{
bic_position(track, sigma, dims = c("x", "y"), m = NULL, log_lik = FALSE)
}
\arguments{
\item{track}{a single track, e.g. an element of a tracks or trax object.}

\item{sigma}{SD of the Gaussian distribution.}

\item{dims}{dimensions used for computation of the statistic.}

\item{m}{the point, expressed as a step number used for splitting the tracks.
The Gaussian mode is fitted to each track subset.}

\item{log_lik}{logical: if TRUE, log likelihood is returned, otherwise
the BIC value.}
}
\value{
a numeric value.
}
\description{
A simple motility test. Positions of a cell without any active
motility (e.g. subjected to Brownian motion) are expected to follow a
multi-variate Gaussian distribution around some central point with a
sigma standard deviation. If the cell moves actively, the displacements are
expected to be a sum of the Gaussian noise and displacement vectors. Hence,
more than just one Gaussian is needed to model it's movement. The function
tests, based on the Bayesian Information Criterion (BIC), the goodness of fit
of the one or two Gaussian model.
}
\details{
The idea behind it comes from the celltracR package vignette:
https://cran.rstudio.com/web/packages/celltrackR/vignettes/QC.html.
}
