% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{c.trax}
\alias{c.trax}
\title{Concatenate two or more trax objects.}
\usage{
\method{c}{trax}(...)
}
\arguments{
\item{...}{one or more trax objects.}
}
\value{
a trax object.
}
\description{
Concatenates trax object into a larger trax object.
}
