% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{calculate.trax}
\alias{calculate.trax}
\title{Calculate track measures.}
\usage{
\method{calculate}{trax}(
  x,
  type = c("steps", "lengths", "times", "duration", "displacements", "speeds",
    "overall_angles", "mean_angles", "overall_dots", "normal_dots", "straightness",
    "asphericity", "delta_BIC", "autocov", "hotellings", "cell_pairs", "step_pairs"),
  ...
)
}
\arguments{
\item{x}{a trax object.}

\item{type}{statistic type, see: Details. 'steps' by default.}

\item{...}{additional arguments to statiscit computing function, see: Details.}
}
\value{
a tibble with the requested statistic, track id
and optionally step id.
}
\description{
Calculates the requested track statistic for a trax object.
}
\details{
Types of returned statistics:
'steps' (default): number of steps per track (\code{\link{track_steps}}),
'lengths': track length (\code{\link{track_lengths}}),
'times': track's step time intervals (\code{\link{track_times}}),
'duration': track duration (\code{\link{track_duration}}),
'displacements': track displacements (\code{\link{track_displacements}}),
'speeds': track speeds (\code{\link{track_speeds}}),
'overall_angles': track overall angles (\code{\link{track_overall_angles}}),
'mean_angles': track mean turning angles (\code{\link{track_mean_angles}}),
'overall_dots': overall dot products (\code{\link{track_overall_dots}})
'normal_dots': normal dot products (\code{\link{track_normal_dots}}),
'straightness': track straightness (\code{\link{track_straightness}}),
'asphericity: track asphericity (\code{\link{track_asphericity}}),
'delta_BIC': track's elements Gaussian modeling outcome
(\code{\link{track_bic_motility}}),
'autocov': autocovariance (\code{\link{autocov}})
'hotellings': results of hotellings test (\code{\link{hotellings_test}}),
'cell_pairs': cell painr distance and angles (\code{\link{analyzeCellPairs}}),
'step_pairs': step pair distance and angles (\code{\link{analyzeStepPairs}}).
}
