% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deltaBIC}
\alias{deltaBIC}
\title{Check if cell motion can be described by multiple Gaussian distributions.}
\usage{
deltaBIC(track, sigma, dims = c("x", "y"))
}
\arguments{
\item{track}{a single track, e.g. an element of a tracks or trax object.}

\item{sigma}{SD of the Gaussian distribution.}

\item{dims}{dimensions used for computation of the statistic.}
}
\value{
a numeric value being the difference in BIC between the one- and
two - Gaussian model.
}
\description{
Checks if the cell motility may be described by one or two
Gaussian models by computing the difference in Bayesian Information
Criterion (BIC).
}
\details{
See: \code{\link{bic_position}}.
}
