% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{filter.trax}
\alias{filter.trax}
\title{Filter a trax object content using the tidyverse API.}
\usage{
\method{filter}{trax}(.data, ..., .preserve = FALSE, rm_empty = TRUE)
}
\arguments{
\item{.data}{a trax object.}

\item{...}{expressions that return a logical value used for the record
selection.}

\item{.preserve}{logical, ignored.}

\item{rm_empty}{logical, should empty tracks be removed?}
}
\value{
a trax object.
}
\description{
Filters a trax object by logical criteria applied to 'id', 'x',
'y', 'z' or 't' parameters.
}
