% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_funs.R
\name{filter_motility}
\alias{filter_motility}
\title{Filter out tracks for non-motile cells.}
\usage{
filter_motility(
  x,
  bic_cutoff = 6,
  sigma = 10,
  dims = c("x", "y"),
  return_both = FALSE
)
}
\arguments{
\item{x}{a trax object.}

\item{bic_cutoff}{BIC (Bayesian Information Criterion) cutoff for
identification of motile cells (cells with >= cutoff are kept).
The value of 6 is a good starting point.}

\item{sigma}{SD of the Gaussian distribution. The expected cell diameter
is usually a good approximation.}

\item{dims}{dimensions used for computation of the statistic.}

\item{return_both}{logical: if TRUE, both the cells passing and missing
the filtering procedure are returned. FALSE by default.}
}
\value{
a trax object or, if return_both is TRUE, a list of trax object with
the 'outcome' (cells passing the procedure) and 'miss' elements, each of them
is a trax object.
}
\description{
Filters out non-motile cells based on the results of
Gaussian modeling.
}
\details{
See: \code{\link{bic_position}}, \code{\link{detlaBIC}} and
https://cran.rstudio.com/web/packages/celltrackR/vignettes/QC.html.
}
