% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_funs.R
\name{filter_multiplets}
\alias{filter_multiplets}
\title{Remove multiplets.}
\usage{
filter_multiplets(
  x,
  method = c("cells", "steps"),
  angle_cutoff = 10,
  dist_cutoff = 10,
  return_both = FALSE,
  ...
)
}
\arguments{
\item{x}{a trax object.}

\item{method}{analysis method, 'cells' (default) or 'steps'.}

\item{angle_cutoff}{angle cutoff. By default 10 degrees
(pairs >= cutoff are kept).}

\item{dist_cutoff}{distance cutoff (pairs >= cutoff are kept).
An approximate cell diameter is a good starting value.}

\item{return_both}{logical: if TRUE, both the cells passing and missing
the filtering procedure are returned. FALSE by default.}

\item{...}{extra arguments passed to
\code{\link[celltrackR]{analyzeCellPairs}} or
\code{\link[celltrackR]{analyzeStepPairs}}.}
}
\value{
a trax object or, if return_both is TRUE, a list of trax object with
the 'outcome' (cells passing the procedure) and 'miss' elements, each of them
is a trax object.
}
\description{
Removes candidate multiplets from the trax objects based on
results of the cell pair analysis. Potential multiplets are characterized
by small distance and low angle.
}
\details{
The pair statistics is calculated with
\code{\link[celltrackR]{analyzeCellPairs}} and
\code{\link[celltrackR]{analyzeStepPairs}}.
}
