% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_funs.R
\name{filter_steps}
\alias{filter_steps}
\title{Filter tracks by step number range.}
\usage{
filter_steps(x, min_steps = NULL, max_steps = NULL, return_both = FALSE)
}
\arguments{
\item{x}{a trax object.}

\item{min_steps}{minimal step number of the track to be kept.
Ignored, if NULL (default).}

\item{max_steps}{maximal step number of the track to be kept.
Ignored, if NULL (default).}

\item{return_both}{logical: if TRUE, both the cells passing and missing
the filtering procedure are returned. FALSE by default.}
}
\value{
a trax object or, if return_both is TRUE, a list of trax object with
the 'outcome' (cells passing the procedure) and 'miss' elements, each of them
is a trax object.
}
\description{
Filters a trax object with the user-provided step number range.
}
