% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{get_mean_speed}
\alias{get_mean_speed}
\title{Calculate mean speed vector.}
\usage{
get_mean_speed(x)
}
\arguments{
\item{x}{a trax object with a single track representing
the sum displacement vector.}
}
\value{
a vector x, y and z dimensions.
}
\description{
Calculates mean speed vector by dividing the mean total
displacement vector by median time interval.
}
