% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{get_total_vectors}
\alias{get_total_vectors}
\title{Calculate total displacement vectors}
\usage{
get_total_vectors(x, simplify = FALSE)
}
\arguments{
\item{x}{a trax object.}

\item{simplify}{logical, should the vector be displayed as a unit vector?
(x, y, z)?}
}
\value{
a trax object or if simplify is TRUE, a tibble with with
the vector dimensions.
}
\description{
Calculates the total displacement vectors for each track.
}
