% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_color_hue_}
\alias{gg_color_hue_}
\title{Emulate the default ggplot2 color palette.}
\usage{
gg_color_hue_(n)
}
\arguments{
\item{n}{number of subsequent colors.}
}
\value{
a character vector with color names.
}
\description{
Choose n subsequent colors from the defult ggplot2 palette,
}
\details{
The idea was presented at
https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
