% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{hotellings_test}
\alias{hotellings_test}
\title{Hotellings test.}
\usage{
hotellings_test(x, dims = c("x", "y"), step_spacing = 0)
}
\arguments{
\item{x}{a trax object.}

\item{dims}{dimensions to be analyzed.}

\item{step_spacing}{how many positions are to be left out between the steps
that are considered for the test,
see: \code{\link[celltrackR]{hotellingsTest}} for details.}
}
\value{
a tibble with statistic value, degrees of freedom and p value.
}
\description{
A pipeline-friendly version of the hotellings test implemented
by \code{\link[celltrackR]{hotellingsTest}}.
}
