% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is_trax}
\alias{is_trax}
\title{Tests for the 'trax' class.}
\usage{
is_trax(x, ...)
}
\arguments{
\item{x}{an object.}

\item{....}{extra arguments, currently none.}
}
\value{
a logical value.
}
\description{
Tests if the object is an instance of the 'trax' class.
}
