% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{plot.trax}
\alias{plot.trax}
\title{Plot the trax object tracks, hotellings vectors or statistics.}
\usage{
\method{plot}{trax}(
  x,
  type = c("tracks", "vectors", "hotellings", "autocov", "statistic", "stat_pair",
    "pair_analysis"),
  normalize = FALSE,
  coverage = 1,
  plot_title = NULL,
  plot_subtitle = NULL,
  cust_theme = theme_trax(),
  ...
)
}
\arguments{
\item{x}{a trax object.}

\item{type}{plot type, see: Details. 'tracks' by default.}

\item{normalize}{logica: should the tracks be normalized prior to plotting?}

\item{coverage}{fraction of the tracks to be presented in the plot. If 1 all
tracks are presented. If < 1, a random sample of tracks of the given size
is presented.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot_subtitle.}

\item{cust_theme}{custom ggplot2 theme.}

\item{...}{extra arguments passed to downstream functions, see: Details.}
}
\description{
Generates a requested plot type for the trax object.
}
\details{
Types and corresponding functions:
'tracks' (default): plots single tracks as specified
in \code{\link{plot_tracks}},
'vectors': plots total displacement vectors with \code{\link{plot_vectors}},
'hotellings': generates a hotellings plot with \code{\link{plot_hotellings}},
'autocov': autocovariance plot with \code{\link{plot_autocov}},
'statistic': tracks' statistic distribution with \code{\link{plot_statistic}},
'stat_pair': tracks' statistic pair as a point plot
with \code{\link{plot_stat_pair}},
'pair_analysis': cell or step pair distances and angles
as specified in \code{\link{plot_pair_analysis}}.
}
