% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{plot_autocov}
\alias{plot_autocov}
\title{Draw autocovariance of displacement vectors.}
\usage{
plot_autocov(
  x,
  method = c("dot_product", "angle"),
  coverage = 1,
  color = NULL,
  mean_color = "orangered3",
  mean_size = 0.75,
  line_size = 0.5,
  line_alpha = 0.5,
  point_size = 2,
  seed = NULL
)
}
\arguments{
\item{x}{a trax object.}

\item{method}{the method of autocovariance calculation: 'dot_product'
(default) or 'angle', as descibed in details.}

\item{coverage}{fraction of the tracks to be presented in the plot. If 1 all
tracks are presented. If < 1, a random sample of tracks of the given size
is presented.}

\item{color}{color name or value for the single tracks' covariances.
If NULL (default), single covariance paths are not shown.}

\item{mean_color}{color of the mean vector autocovariance line.}

\item{mean_size}{width of the mean vector autocovariance line.}

\item{line_size}{line width for single track autocovariances.}

\item{line_alpha}{line alpha for single track autocovariances.}

\item{point_size}{size of the points representing the mean autocovariances.}

\item{seed}{seed value for random number generation. Ignored if coverage = 1.}
}
\value{
a ggplot2 object.
}
\description{
Plots step-wise values of autocovariance of displacement
vectors.
}
\details{
The autocovariance is calculated as a dot product of the vectors
returned by \code{\link[celltrackR]{overallDot}} or overall angle computed
by \code{\link[celltrackR]{overallAngle}}
}
