% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{plot_hotellings}
\alias{plot_hotellings}
\title{Draw a 2D hotellings plot.}
\usage{
plot_hotellings(
  x,
  dims = c("x", "y"),
  step_spacing = 0,
  conf_level = 0.95,
  point_alpha = 0.25,
  point_color = "steelblue",
  ellipse_border = "coral4",
  ellipse_color = "coral3",
  ellipse_alpha = 0.25,
  center_color = "coral4",
  center_shape = 16,
  center_size = 3
)
}
\arguments{
\item{x}{a trax object.}

\item{dims}{a pair of dimensions to be plotted.}

\item{step_spacing}{how many positions are to be left out between the steps
that are considered for the test,
see: \code{\link[celltrackR]{hotellingsTest}} for details.}

\item{conf_level}{confidence level.}

\item{point_alpha}{alpha of points.}

\item{point_color}{point color.}

\item{ellipse_border}{ellipse border color.}

\item{ellipse_color}{ellipse fill color.}

\item{ellipse_alpha}{ellipse alpha}

\item{center_color}{ellipse center color.}

\item{center_shape}{shape of the ellipse center point.}

\item{center_size}{size of the ellipse center point.}
}
\description{
Draws a two-dimensional hotellings plot with the confidence
interval ellipsis and mean displacement.
}
\details{
a wrapper around \code{\link[celltrackR]{hotellingsTest}}
}
