% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{plot_pair_analysis}
\alias{plot_pair_analysis}
\title{Plot distance and/or angle between cell or step pairs.}
\usage{
plot_pair_analysis(
  x,
  method = c("cells", "steps"),
  stat = c("both", "dist", "angle"),
  geom = c("histogram", "violin", "boxpplot"),
  coverage = 1,
  trend_method = "none",
  color = "steelblue",
  point_alpha = 0.25,
  point_size = 2,
  point_jitter = 0.1,
  shape_color = color,
  shape_alpha = 0.25,
  line_size = 0.75,
  line_color = "orangered3",
  line_alpha = 0.5,
  analysis_args = rlang::list2(),
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{a cell trax object.}

\item{method}{analysis method, 'cells' (default) or 'steps'.}

\item{stat}{statistic to be shown in the plot: 'both' displays the angles
and distances, 'angle' or 'dist' produce distribution plots for
single statistics.}

\item{geom}{distribution plot form: 'histogram' (default), 'violin'
or 'boxplot'. Ignored if stat argument is 'both'.}

\item{coverage}{fraction of the tracks to be presented in the plot. If 1 all
tracks are presented. If < 1, a random sample of tracks of the given size
is presented.}

\item{trend_method}{method/function for the trend line fitting, as described
for the 'method' argument of \code{\link[ggplot2]{geom_smooth}}.
If 'none' (default), no trend line is plotted.}

\item{color}{point color.}

\item{point_alpha}{point alpha.}

\item{point_size}{poin size.}

\item{point_jitter}{jitter width for data points in the distribution plots,
ignored if stat is 'both'.}

\item{shape_color}{color of the shape in the distribution plot.}

\item{shape_alpha}{alpha of the shape in the distribution plot.}

\item{line_color}{trend line color.}

\item{line_alpha}{alpha for the trend confidence interval.}

\item{analysis_args}{additional arguments passed to analysis functions
\code{\link[celltrackR]{analyzeCellPairs}} or
\code{\link[celltrackR]{analyzeStepPairs}}}

\item{seed}{seed value for random number generation. Ignored if coverage = 1.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_smooth}},
such as sd, step and so on in case, when both stats are plotted. For
distribution plots, extra arguments passed to
\code{\link[ggplot2]{geom_histogram}}, \code{\link[ggplot2]{geom_violin}} or
\code{\link[ggplot2]{geom_boxplot}}.}
}
\value{
a ggplot object.
}
\description{
Plots distribution of angle or distance or values of
angle versus distance between the cell or step pairs.
}
