% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{plot_stat_pair}
\alias{plot_stat_pair}
\title{Plot a pair of track measures.}
\usage{
plot_stat_pair(
  x,
  x_stat = "mean_angles",
  y_stat = "normal_dots",
  coverage = 1,
  color = "steelblue",
  point_size = 2,
  point_alpha = 0.5,
  point_jitter = 0,
  seed = NULL,
  x_args = rlang::list2(),
  y_args = rlang::list2()
)
}
\arguments{
\item{x}{a trax object.}

\item{x_stat}{statistic to be presented on the X axis, see Details.}

\item{y_stat}{statistic to be presented on the Y axis, see Details.}

\item{coverage}{fraction of the tracks to be presented in the plot. If 1 all
tracks are presented. If < 1, a random sample of tracks of the given size
is presented.}

\item{color}{point color.}

\item{point_size}{point size.}

\item{point_alpha}{point alpha.}

\item{point_jitter}{point jitter width.}

\item{seed}{seed value for random number generation. Ignored if coverage = 1.}

\item{x_args}{extra arguments to the X statistic calculation functions as
specified in Details. Provided as a named list.}

\item{y_args}{extra arguments to the Y statistic calculation functions as
specified in Details. Provided as a named list.}
}
\value{
a ggplot2 object.
}
\description{
Generates a point plot with a pair of track statistics,
see: Details.
}
\details{
'steps' (default): number of steps per track (\code{\link{track_steps}}),
'lengths': track length (\code{\link{track_lengths}}),
'times': track's step time intervals (\code{\link{track_times}}),
'duration': track duration (\code{\link{track_duration}}),
'displacements': track displacements (\code{\link{track_displacements}}),
'speeds': track speeds (\code{\link{track_speeds}}),
'overall_angles': track overall angles (\code{\link{track_overall_angles}}),
'mean_angles': track mean turning angles (\code{\link{track_mean_angles}}),
'overall_dots': overall dot products (\code{\link{track_overall_dots}})
'normal_dots': normal dot products (\code{\link{track_normal_dots}}),
'straightness': track straightness (\code{\link{track_straightness}}),
'asphericity: track asphericity (\code{\link{track_asphericity}}),
'delta_BIC': track's elements Gaussian modeling outcome
(\code{\link{track_bic_motility}}).
}
