% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{plot_statistic}
\alias{plot_statistic}
\title{Plot track measure.}
\usage{
plot_statistic(
  x,
  stat = c("steps", "lengths", "times", "duration", "displacements", "speeds",
    "overall_angles", "mean_angles", "overall_dots", "normal_dots", "straightness",
    "asphericity", "delta_BIC"),
  geom = c("histogram", "violin", "boxplot"),
  color = "steelblue",
  point_color = color,
  point_size = 2,
  point_alpha = 0.5,
  point_jitter = 0.1,
  shape_alpha = 0.25,
  bins = NULL,
  scale = "width",
  draw_quantiles = c(0.25, 0.5, 0.75),
  ...
)
}
\arguments{
\item{x}{a trax object.}

\item{stat}{statistic name as specified in Details,}

\item{geom}{plot form: 'histogram' (default), 'violin' or 'boxplot'.}

\item{color}{shape color.}

\item{point_color}{point color.}

\item{point_size}{point size.}

\item{point_alpha}{point alpha.}

\item{point_jitter}{point jitter width.}

\item{shape_alpha}{shape alpha.}

\item{bins}{number of histogram bins.}

\item{scale}{violin plot area/width scaling (\code{\link[ggplot2]{geom_violin}}).}

\item{draw_quantiles}{violin plot quantiles (\code{\link[ggplot2]{geom_violin}}).}

\item{...}{extra arguments to the statistic calculation functions as
specified in Details.}

\item{seed}{seed value for random number generation. Ignored if coverage = 1.}
}
\value{
a ggplot2 object.
}
\description{
Plots distribution of various track statistics as histograms,
violin or box plots.
}
\details{
'steps' (default): number of steps per track (\code{\link{track_steps}}),
'lengths': track length (\code{\link{track_lengths}}),
'times': track's step time intervals (\code{\link{track_times}}),
'duration': track duration (\code{\link{track_duration}}),
'displacements': track displacements (\code{\link{track_displacements}}),
'speeds': track speeds (\code{\link{track_speeds}}),
'overall_angles': track overall angles (\code{\link{track_overall_angles}}),
'mean_angles': track mean turning angles (\code{\link{track_mean_angles}}),
'overall_dots': overall dot products (\code{\link{track_overall_dots}})
'normal_dots': normal dot products (\code{\link{track_normal_dots}}),
'straightness': track straightness (\code{\link{track_straightness}}),
'asphericity: track asphericity (\code{\link{track_asphericity}}),
'delta_BIC': track's elements Gaussian modeling outcome
(\code{\link{track_bic_motility}}).
}
