% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{plot_tracks}
\alias{plot_tracks}
\title{Draw single tracks.}
\usage{
plot_tracks(
  x,
  dims = c("x", "y"),
  normalize = FALSE,
  coverage = 1,
  color = NULL,
  show_zero = FALSE,
  point_size = 1,
  point_alpha = 1,
  line_size = 0.5,
  line_alpha = 1,
  end_size = 4,
  legend = "none",
  seed = NULL
)
}
\arguments{
\item{x}{a trax object.}

\item{dims}{a pair of dimensions to be plotted.}

\item{normalize}{logical, should the tracks be normalized?}

\item{coverage}{fraction of the tracks to be presented in the plot. If 1 all
tracks are presented. If < 1, a random sample of tracks of the given size
is presented.}

\item{color}{color name or value. If NULL (default), each path is colored
in a default ggplot2 style and can be modified by adjusting manual color
scale, see: \code{\link[ggplot2]{scale_color_manual}} for details.}

\item{show_zero}{logical, should the X = 0 and Y = 0 lines be plotted?}

\item{point_size}{point size.}

\item{point_alpha}{point alpha.}

\item{line_size}{line width.}

\item{line_alpha}{line alpha.}

\item{end_size}{endpoint size.}

\item{legend}{form of the color legend, as specified in
\code{\link[ggplot2]{guides}}. Defaults to 'none'.}

\item{seed}{seed value for random number generation. Ignored if coverage = 1.}
}
\value{
a ggplot2 object.
}
\description{
Draws single tracks given a trax object.
}
