% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_utils.R
\name{plot_vectors}
\alias{plot_vectors}
\title{Draw total displacement vectors.}
\usage{
plot_vectors(
  x,
  dims = c("x", "y"),
  normalize = FALSE,
  coverage = 1,
  color = NULL,
  mean_color = "orangered3",
  mean_size = 0.75,
  show_zero = TRUE,
  line_size = 0.5,
  line_alpha = 1,
  arrow = grid::arrow(angle = 30, length = grid::unit(3, "points"), ends = "last", type
    = "open"),
  mean_arrow = grid::arrow(angle = 45, length = grid::unit(12, "points"), ends =
    "last", type = "open"),
  legend = "none",
  seed = NULL
)
}
\arguments{
\item{x}{a trax object.}

\item{dims}{a pair of dimensions to be plotted.}

\item{normalize}{logical, should the tracks be normalized?}

\item{coverage}{fraction of the tracks to be presented in the plot. If 1 all
tracks are presented. If < 1, a random sample of tracks of the given size
is presented.}

\item{color}{color name or value. If NULL (default), each path is colored
in a default ggplot2 style and can be modified by adjusting manual color
scale, see: \code{\link[ggplot2]{scale_color_manual}} for details.}

\item{mean_color}{color of the mean vector. If NULL, no mean vector is
displayed. Ignored if normalize = FALSE.}

\item{mean_size}{width of the mean vector line.}

\item{show_zero}{logical, should the X = 0 and Y = 0 lines be plotted?}

\item{line_size}{line width.}

\item{line_alpha}{line alpha.}

\item{arrow}{arrow specification for the mean vector,
as created by \code{\link[grid]{arrow}}.}

\item{legend}{form of the color legend, as specified in
\code{\link[ggplot2]{guides}}. Defaults to 'none'.}

\item{seed}{seed value for random number generation. Ignored if coverage = 1.}
}
\value{
a ggplot2 object.
}
\description{
Draws total displacement vectors for each track.
}
