% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{print.trax}
\alias{print.trax}
\title{Print a trax object heading.}
\usage{
\method{print}{trax}(x, ...)
}
\arguments{
\item{x}{a trax object.}

\item{...}{extra arguments passed to methods.}
}
\description{
Prints first names of the trax object tracks, time step and
the total number of tracks.
}
