% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_funs.R
\name{read_trax}
\alias{read_trax}
\title{Imports tracks from a single text file.}
\usage{
read_trax(file, ...)
}
\arguments{
\item{file}{a path to the file.}

\item{...}{extra arguments passed to \code{\link[readr]{read_tsv}}.}
}
\description{
Reads track information from a tab-separated text file of the
following columns named: id, t, x, y and optionally z, containing track id,
time, x, y and optionally z coordinates, respectively.
}
\details{
Note: tracks with less than 2 steps are automatically removed.
}
