% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_funs.R
\name{read_trax_parts}
\alias{read_trax_parts}
\title{Import of track data from separate x, y, z position text files.}
\usage{
read_trax_parts(x_file, y_file, z_file = NULL, ...)
}
\arguments{
\item{x_file}{a path to the file containing X coordinates.}

\item{y_file}{a path to the file containing Y coordinates.}

\item{z_file}{a path to the file containing Z coordinates. Optional, if not
provided, the tracks will be constrained to the X and Y coordinates.}

\item{...}{extra arguments passed to \code{\link[readr]{read_tsv}}.}
}
\value{
a trax class object.
}
\description{
The function enables for import of track data from separate
tab-separated text files containing the X, Y and Z positions, respectively.
}
\details{
The cells are identified by consistent heading of each file.
The obligatory file format requires time in the first column and track names
in the remaining headings and coordinate values distributed in rows. Note:
tracks with less than 2 steps are automatically removed.
}
