% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_funs.R
\name{read_trax_text}
\alias{read_trax_text}
\title{Import tracks from a single unformatted text file.}
\usage{
read_trax_text(
  file,
  id_name,
  t_name,
  x_name,
  y_name,
  z_name = NULL,
  sample_name = NULL,
  delim = "\\t",
  ...
)
}
\arguments{
\item{file}{a path to the text file.}

\item{id_name}{name of the column storing track IDs.}

\item{t_name}{name of the column storing times.}

\item{x_name}{name of the column storing x coordinates.}

\item{y_name}{name of the column storing y coordinates.}

\item{z_name}{name of the column storing z coordinates, optional.}

\item{sample_name}{name of the column storing sample IDs, optional.}

\item{...}{extra arguments passed to \code{\link[readr]{read_delim}}.}
}
\value{
a trax class object or, if multiple samples defined (sample_id),
a list of trax objects named after the provided sample IDs.
}
\description{
This the most flexible way of importing tracks from a text file.
It handles most of tabular text formats with named columns. th user is asked
to specify names of columns containing track ID, time and coordinates. It may
work considerably slower than \code{\link{read_trax}} or
\code{\link{read_trax_parts}}.
}
\details{
Note: tracks with less than 2 steps are automatically removed. Uses
\code{\link[readr]{read_delim}} under the hood.
}
