% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{sample_trax}
\alias{sample_trax}
\title{Sample tracks at random.}
\usage{
sample_trax(x, size, replace = FALSE, seed = NULL)
}
\arguments{
\item{x}{a tracks object.}

\item{size}{sample size.}

\item{replace}{logical: should the sampling be done with replacement?}

\item{seed}{seed value for random number generation.}
}
\value{
a trax object.
}
\description{
Gets a random sample of tracks.
}
