% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_colors_}
\alias{set_colors_}
\title{Choose colors randomly from the grDevice palette.}
\usage{
set_colors_(n, seed = NULL)
}
\arguments{
\item{n}{number of colors to choose.}

\item{seed}{seed for the random number generator.}
}
\value{
a character vector with color names.
}
\description{
Chooses n colors at random from the palette provided by
\code{\link[grDevices]{colors}}.
}
