% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_track}
\alias{split_track}
\title{Split a single track by a displacement cutoff.}
\usage{
split_track(track, disp_cutoff, prefix = NULL)
}
\arguments{
\item{track}{a single track, i.e. element of tracks or trax object.}

\item{disp_cutoff}{the scalar displacement cutoff.}

\item{prefix}{a prefix of the track split names.}
}
\value{
a trax object.
}
\description{
Given a single track, the function calculated displacements for
each step and splits the track into multiple ones if the cutoff is met.
}
\details{
Returns NULL, if the track has only displacements larger than
the disp_cutoff parameter.
}
