% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_funs.R
\name{split_tracks}
\alias{split_tracks}
\title{Split tracks by a displacement cutoff.}
\usage{
split_tracks(x, disp_cutoff, return_both = FALSE)
}
\arguments{
\item{x}{a trax object.}

\item{disp_cutoff}{the scalar displacement cutoff.}

\item{return_both}{logical: if TRUE, both the cells before and after the
transformation. FALSE by default.}
}
\value{
a trax object.
}
\description{
Given a trax object, the function calculates displacements for
each step and splits of the signle track into multiple ones
if the cutoff is met.
}
