% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{[.trax}
\alias{[.trax}
\title{Subset a trax object.}
\usage{
\method{[}{trax}(x, y)
}
\arguments{
\item{x}{a trax object.}

\item{y}{an index or index vector.}
}
\value{
a trax object.
}
\description{
Subsets a trax object in a list-like manner.
}
