% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{summary.trax}
\alias{summary.trax}
\title{Summary of the trax object track measures.}
\usage{
\method{summary}{trax}(object, ...)
}
\arguments{
\item{object}{a trax object.}

\item{...}{additional arguments passed to \code{\link{track_bic_motility}}}
}
\value{
A tibble with the statistic values for each track.
}
\description{
Calculates key track-wise statistics, see Details.
}
\details{
stat names are as follows:
'steps': number of steps per track (\code{\link{track_steps}}),
'length': track length (\code{\link{track_lengths}}),
'duration': track duration (\code{\link{track_duration}}),
'delta_BIC': Gaussian motility test (\code{\link{track_bic_motility}}),
'total_displacement': total track displacements (\code{\link{track_displacements}}),
'mean_displacement': mean track displacements (\code{\link{track_displacements}}),
'median_displacement': median track displacements (\code{\link{track_displacements}}),
'mean_speed': track mean speeds (\code{\link{track_speeds}}),
'median_speed': track median speeds (\code{\link{track_speeds}}),
'overall_angle': track overall angles (\code{\link{track_overall_angles}}),
'mean_angle': track mean turning angles (\code{\link{track_mean_angles}}),
'overall_dot': overall dot products (\code{\link{track_overall_dots}})
'normal_dot': normal dot products (\code{\link{track_normal_dots}}),
'straightness': track straightness (\code{\link{track_straightness}}),
'asphericity: track asphericity (\code{\link{track_asphericity}})),
'x_displ', 'y_displ', 'z_displ': total displacements for each track
(\code{\link{get_total_vectors}}).
}
