% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_funs.R
\name{time_intervals}
\alias{time_intervals}
\title{Detect step time abnormalities.}
\usage{
time_intervals(x)
}
\arguments{
\item{x}{a trax object.}
}
\value{
A tibble with frequency of time intervals.
}
\description{
Detects steps in a trax object with differing time differences.
}
