% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_asphericity}
\alias{track_asphericity}
\title{Calculate track asphericity.}
\usage{
track_asphericity(x)
}
\arguments{
\item{x}{a trax object.}
}
\description{
Calculates track asphericity as described in
\code{\link[celltrackR]{TrackMeasures}}
}
