% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_bic_motility}
\alias{track_bic_motility}
\title{Test for active cell mobility.}
\usage{
track_bic_motility(x, sigma = 10, dims = c("x", "y"))
}
\arguments{
\item{x}{a trax object.}

\item{sigma}{SD of the Gaussian distribution. The approximate cell diameter
is a good starting point here.}

\item{dims}{dimensions used for computation of the statistic.}
}
\value{
A tibble with the delta BIC values (one- vs two-Gaussian model).
}
\description{
The movement of cells without active motility is expected to be
well described by a simple Gaussian model. If the motility is active,
at least two Gaussian distributions are required to accomplish that.
The function compares the one - versus two - Gaussian models applied to
each track by calculating difference in BIC (Bayesian Information Criterion).
}
\details{
See: \code{\link{bic_position}}, \code{\link{detlaBIC}} and
https://cran.rstudio.com/web/packages/celltrackR/vignettes/QC.html. The
function can calculate the delta BIC statistic for tracks with 4 steps or more,
others are ignored and a warning raised.
}
