% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_lengths}
\alias{track_lengths}
\title{Calculate track lengths.}
\usage{
track_lengths(x)
}
\arguments{
\item{x}{a trax object.}
}
\value{
a tibble with lengths for each track.
}
\description{
Returns track lengths.
}
