% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_mean_angles}
\alias{track_mean_angles}
\title{Calculate track mean angle.}
\usage{
track_mean_angles(x, ...)
}
\arguments{
\item{x}{a trax object.}

\item{...}{other arguments passed to
\code{\link[celltrackR]{meanTurningAngle}}}
}
\description{
Calculates track mean turning angle as described in
\code{\link[celltrackR]{TrackMeasures}}
}
