% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_overall_angles}
\alias{track_overall_angles}
\title{Calculate track overall angle.}
\usage{
track_overall_angles(x, ...)
}
\arguments{
\item{x}{a trax object.}

\item{...}{other arguments passed to \code{\link[celltrackR]{overallAngle}}}
}
\description{
Calculates track overall angle as described in
\code{\link[celltrackR]{TrackMeasures}}
}
