% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_overall_dots}
\alias{track_overall_dots}
\title{Calculate overall dot product of displacement vectors.}
\usage{
track_overall_dots(x, ...)
}
\arguments{
\item{x}{a trax object.}

\item{...}{other arguments passed to \code{\link[celltrackR]{overallDot}}}
}
\description{
Calculates track overall dot product of displacement vectors
as described in \code{\link[celltrackR]{TrackMeasures}}
}
