% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_speeds}
\alias{track_speeds}
\title{Calculate scalar speeds.}
\usage{
track_speeds(x, aggregate = c("none", "total", "mean", "median"))
}
\arguments{
\item{x}{a trax object.}

\item{aggregate}{'none' (default): speeds for each step are returned,
'mean' or 'median': mean or median speeds per track are returned, 'total':
total displacement by time.}
}
\value{
a tibble with the requested values.
}
\description{
Calculates speeds, step-wise or track-wise.
}
