% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_steps}
\alias{track_steps}
\title{Calculate track step numbers.}
\usage{
track_steps(x)
}
\arguments{
\item{x}{a trax object.}
}
\value{
a tibble with step numbers for each track.
}
\description{
Returns track step numbers.
}
