% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_straightness}
\alias{track_straightness}
\title{Calculate track straightness.}
\usage{
track_straightness(x)
}
\arguments{
\item{x}{a trax object.}
}
\description{
Calculates track straightness as described in
\code{\link[celltrackR]{TrackMeasures}}
}
