% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funs.R
\name{track_times}
\alias{track_times}
\title{Calculate track time steps.}
\usage{
track_times(x)
}
\arguments{
\item{x}{a trax object.}
}
\value{
a tibble with track duration for each track step.
}
\description{
Returns time intervals for every track step.
}
