% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_funs.R
\name{write_trax}
\alias{write_trax}
\title{Write a trax object locally.}
\usage{
write_trax(x, file, ...)
}
\arguments{
\item{x}{a trax object.}

\item{file}{a path to the file.}

\item{...}{extra arguments passed to \code{\link[readr]{write_tsv}}.}
}
\description{
Writes a trax object locally as a tab-separated text file with
the following columns named: id, t, x, y and optionally z, containing
track id, time, x, y and optionally z coordinates, respectively.
}
